refactor(web): 로그인 배너 불필요한 prop 제거
